+++ package.json
@@ -51,6 +51,7 @@
     "hast-util-to-string": "^1.0.4",
     "husky": "4.3.8",
     "import-fresh": "^3.3.0",
+    "jotai": "^0.16.4",
     "katex": "^0.13.2",
     "loader-utils": "^2.0.0",
     "loadjs": "^4.2.0",
